Move judgment model to separate file
This provides clearer isolation of concerns and keeps files smaller and more manageable, as well as reducing blast radius of changes (or at least giving us a clear sign when our changes have influence in places we wouldn't expect).
