+++ judgments/tests/test_unlock.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 from django.test import Client
 from django.urls import reverse
 
-from judgments.models import Judgment
+from judgments.models.judgments import Judgment
 
 
 @pytest.mark.django_db
